requirements.txt
# Core PyTorch packages
torch>=1.9.0
torchvision>=0.10.0

# Image processing
Pillow>=8.0.0

# Machine learning utilities
scikit-learn>=1.0.0
numpy>=1.21.0